feat(retrieval): implement GraphRAG hybrid retrieval (ADR-0070)
Implemented a unified GraphRAG retriever that combines vector similarity
(anchors) with graph traversal (expansion) and blended scoring:
score = sim_weight * cosine + graph_weight * (1/(1+hops)) * path_strength.

Key improvements:
- Core GraphRAG logic in `src/retrieval/graph_rag.rs` using `VecDeque` for BFS.
- Addressed PR feedback: Traversal now revisits nodes if a path with a
  higher graph score (strength / (1 + hops)) is discovered.
- Maintained strict 500-LOC limits per file using extension modules:
  `src/framework_graph_rag.rs` and `src/wasm_graph_rag.rs`.
- Integrated `csm probe-graph` CLI subcommand.
- Comprehensive integration tests in `tests/graph_rag.rs`.
- Performance benchmark verifies p50 latency of ~0.8ms for 1k concepts/5 hops.

Fixed CI failures:
- Resolved linting issues in `tests/graph_rag.rs` (clone-on-copy and float-cmp).
- Fixed LOC gate failure for `src/framework.rs` and `src/wasm_ext.rs` by using extension modules.

Resolves #140.
