+++ llms-full.txt
@@ -3013,8 +3013,6 @@ Batch similarity computation tuned for cache locality and Rayon overhead.
 
 ## src/hyperdim_simd.rs
 
-## src/hyperdim_tests.rs
-
 ## src/lib.rs
 
 ### ConceptVersion
